backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
redis==5.0.1
anthropic==0.18.1
openai==1.12.0
google-generativeai==0.3.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0